# Organization/Organization.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organization.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/VoucherType": {
      "get": {
        "tags": [
          "VoucherType"
        ],
        "summary": "The method returns all the voucher types in the organization",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetVoucherTypeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetVoucherTypeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetVoucherTypeDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "VoucherType"
        ],
        "summary": "The method adds a new VoucherType into organization",
        "requestBody": {
          "description": "A new VoucherType that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVoucherTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVoucherTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostVoucherTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/VoucherType/{id}": {
      "get": {
        "tags": [
          "VoucherType"
        ],
        "summary": "The method returns an voucher type by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "VoucherType ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "VoucherType"
        ],
        "summary": "The method updates an VoucherType information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the VoucherType",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the VoucherType",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVoucherTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVoucherTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostVoucherTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "VoucherType"
        ],
        "summary": "The method deletes an VoucherType by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the VoucherType",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVoucherTypeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Workshop": {
      "get": {
        "tags": [
          "Workshop"
        ],
        "summary": "The method returns all the workshops in the organization",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetWorkshopDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetWorkshopDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetWorkshopDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workshop"
        ],
        "summary": "The method adds a new workshop into organization",
        "requestBody": {
          "description": "A new workshop that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkshopDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkshopDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkshopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Workshop/{id}": {
      "get": {
        "tags": [
          "Workshop"
        ],
        "summary": "The method returns an workshop by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workshop ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workshop"
        ],
        "summary": "The method updates a workshop information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the workshop",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the workshop",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkshopDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkshopDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkshopDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workshop"
        ],
        "summary": "The method deletes a workshop by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the workshop",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GetVoucherTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - an id of a VoucherType",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name - a name of a VoucherType",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "VoucherType - represents a type of vacation voucher, e.g. \"sanatorium\", \"vacation home\", etc."
      },
      "GetWorkshopDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - an id of the workshop",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Name - a name of the workshop",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class PostWorkshopDto represents a workshop on the organization"
      },
...
      "PostVoucherTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name - a name of a VoucherType",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "VoucherType - represents a type of vacation voucher, e.g. \"sanatorium\", \"vacation home\", etc."
      },
      "PostWorkshopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name - a name of the workshop",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class PostWorkshopDto represents a workshop on the organization"
      }
    }
  }
}